refactor: Password 객체 인터페이스로 분리
