+++ src/main/java/woowacourse/shoppingcart/domain/customer/password/Password.java
@@ -1,0 +1,10 @@
+package woowacourse.shoppingcart.domain.customer.password;
+
+public interface Password {
+
+    void validateMatchingOriginalPassword(final String other);
+
+    void validateMatchingLoginPassword(final String other);
+
+    String getValue();
+}
